requirements.txt
streamlit>=1.28.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
pytest>=7.0.0
